Update api calls to typescript files
